requirements.txt
pyyaml
requests
anthropic